python/pyproject.toml
[project]
name = "zendesk-mcp"
version = "1.0.0"
description = "MCP Server for interacting with the Zendesk API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.25.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "starlette>=0.50.0",
    "uvicorn>=0.40.0",
]

[project.scripts]
zendesk-mcp = "zendesk_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["zendesk_mcp"]
